code/requirements/base.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.10
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
backoff==1.6.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bitcoin==1.1.42
cached-property==1.5.1
cachetools==4.2.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dateparser==0.7.4
defusedxml==0.6.0
Django==3.1.1
django-admin-rangefilter==0.6.3
django-braces==1.13.0
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.1
django-dbbackup==3.3.0
django-debug-toolbar==3.2
django-extensions==3.0.3
django-filter==2.4.0
django-formtools==2.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-rest-passwordreset==1.1.0
django-rest-polymorphic==0.1.8
django-rest-verification==0.42
django-storages==1.9.1
django-tinymce==2.8.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
drf-yasg==1.20.0
dropbox==9.4.0
facebook-sdk==3.1.0
flower==0.9.3
future==0.17.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
isodate==0.6.0
itypes==1.1.0
jet-bridge-base==0.7.6
Jinja2==2.10.1
jsonfield==3.1.0
kiwisolver==1.1.0
kombu==4.6.11
lxml==4.5.0
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.17.4
oauthlib==3.1.0
packaging==20.4
phonenumbers==8.12.11
Pillow==7.1.1
pkginfo==1.6.0
posthog==1.1.3
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycoin==0.90.20190630
PyJWT==1.7.1
pyparsing==2.4.5
python-bidi==0.4.2
python-dateutil==2.8.0
python-gnupg==0.4.5
python-http-client==3.1.0
pytz==2019.1
redis==3.3.7
regex==2020.5.14
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
rtl==0.4.3
ruamel.yaml==0.15.99
sendgrid==6.0.5
sib-api-v3-sdk==6.0.2
simplejson==3.17.0
six==1.12.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.0
tornado==5.1.1
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.9
zeep==3.4.0
zipp==3.0.0
